requirements.txt
arxiv>=2.1.0
feedparser>=6.0.10
google-generativeai>=0.5.0
PyGithub>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
PyMuPDF>=1.23.0
